Update PIR-alarm-system
Now we're going to build the most sophisticated alarm system in the world:)
Our PIR sensor acts like a sonar, alternating between active and passive status. At 2 seconds intervals it checks if there is any movement nearby. 
If any motion gets detected the LED light is turned on and buzzer sounds the alarm, which makes you blind and deaf respectively :))