Update Knight-Rider
Knight Rider :) if you know what it was then you're feeling the melancholy of nostalgia right now. 
We have multiple LEDs lined up straight. Our code makes them turn on and off sequentially along the line. When the light travels from end-to-end it returns back to the other side. Going back and forth in an infinite loop.